reverted refactoring in order to make PR more easy to review
